Add files via upload
Write a program that determines whether a truck is loases beyond its rated capacity. The program will read in the truck's maximum load capacity (in metic tons), the number of boxes it's xarrying and the acerage weight of a box. If the total weight of the boxes is less than or equal to the trucks capacity, the program announces that the truck can legally carry all boxes and tells how many additonal boxes can be addes. If the number of boxes exceeds the maximum capacity, the program announces that the truck is overloaded and tells how many boxes must be removed