Add files via upload
Write a program that reads the speed of an airplane in miles per hour and the travel time in hours. The pogram should dispaly the distance traveled by the airplane in that time in miles, and its speed in meters per second. Use the equation: Speed in m/s = (Speed in MPH / 3600) * 1609.344